feat: 간병사 번호 추가

+++ app/src/main/java/com/ezlevup/ganbyeong24/data/model/CareRequest.kt
@@ -25,7 +25,7 @@ import com.google.firebase.firestore.Exclude
  */
 data class CareRequest(
         @get:Exclude val id: String = "",
-        val serialNumber: Int = 0,
+        val serialNumber: Long = 0,
         val userId: String = "",
         val patientName: String = "",
         val patientAge: Int = 0,
